form change system partial implementation

+++ LetMeRIP/Assets/Scripts/Player/PlayerStats.cs
@@ -10,9 +10,9 @@ public class PlayerStats : ScriptableObject
 	public float maxHealth;
 	public float spiritGauge;
 	public float maxSpiritGauge;
-	public float strength;
-	public float dexterity;
-	public float intelligence;
+	public int strength;
+	public int dexterity;
+	public int intelligence;
 
 	public float defense;
 	public float swiftness;
